+++ src/pupil_labs/realtime_api/base.py
@@ -4,7 +4,7 @@ import typing as T
 from .models import APIPath, DiscoveredDeviceInfo
 
 
-class ControlBase(abc.ABC):
+class DeviceBase(abc.ABC):
     def __init__(
         self,
         address: str,
@@ -28,7 +28,7 @@ class ControlBase(abc.ABC):
         return f"Control(ip={self.address}, port={self.port}, dns={self.dns_name})"
 
     @classmethod
-    def from_discovered_device(cls, device: DiscoveredDeviceInfo) -> "ControlBase":
+    def from_discovered_device(cls, device: DiscoveredDeviceInfo) -> "DeviceBase":
         return cls(
             device.addresses[0],
             device.port,
@@ -37,7 +37,7 @@ class ControlBase(abc.ABC):
         )
 
     @classmethod
-    def convert_from(cls, other: "ControlBase") -> "ControlBase":
+    def convert_from(cls, other: "DeviceBase") -> "DeviceBase":
         return cls(
             other.address,
             other.port,